requirements.txt
mysql-connector-python~=9.5.0

